Change how responses are stored.
Make an explicit type for "Local" responses, which come off the local disk from
a previous cached called and "Remote" responses which are fetched remotely.
Both types can be sent to the client as-is. Remote responses are serialized to
disk, while Local responses are loaded from disk.

The [de]serialization of responses has moved from gob to a custom format. This
was done for future work where local files might be cleared based on their
expiration. The custom format will make this possible without having to read
the entire response from disk (which could include a large-ish body).

In addition to the above work, the concept of expiration has been added to the
configuration.