Don't persist 404s to disk
Saving 404 response objects to disk for caching is too easy an attack vector to
exploint: someone can issue random requests to eat up disk space. 404s are now
treated differently than other responses in that a simple in-memory cache is
used. This has the downside that the actual upstream's 404 response isn't used.